toolbar dropdown button structure update
