+++ src/components/mails-toolbar-display-dropdown-button/index.module.css
@@ -1,0 +1,6 @@
+.wrapper {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
