+++ src/components/mails-toolbar-dropdown-button/index.module.css
@@ -1,6 +1,16 @@
-.wrapper {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
+.button {
+  font-size: 0.7em;
+  background-color: transparent;
+  font-weight: bold;
+  border: none;
+  width: 240px;
+  padding: 10px;
+  margin: 3px 0;
+  letter-spacing: 2px;
+  cursor: pointer;
+}
+
+.button:hover {
+  background-color: #24292e;
+  color: #ffffff;
 }
