--- src/components/mails-toolbar-dropdown-move-to/index.module.css
@@ -1,16 +1,0 @@
-.button {
-  font-size: 0.7em;
-  background-color: transparent;
-  font-weight: bold;
-  border: none;
-  width: 240px;
-  padding: 10px;
-  margin: 3px 0;
-  letter-spacing: 2px;
-  cursor: pointer;
-}
-
-.button:hover {
-  background-color: #24292e;
-  color: #ffffff;
-}
